+++ framework/Furion/DynamicApiController/Conventions/DynamicApiControllerApplicationModelConvention.cs
@@ -653,7 +653,7 @@ namespace Furion.DynamicApiController
         private static void ConfigureActionUnifyResultAttribute(ActionModel action)
         {
             // 判断是否手动添加了标注或跳过规范化处理
-            if (UnifyContext.IsSkipUnifyHandlerOnSucceedReturn(action.ActionMethod, out var _, false)) return;
+            if (UnifyContext.CheckSucceeded(action.ActionMethod, out var _, false)) return;
 
             // 获取真实类型
             var returnType = action.ActionMethod.GetRealReturnType();