+++ framework/Furion/UnifyResult/Options/UnifyResultStatusCodesOptions.cs
@@ -23,9 +23,9 @@ namespace Furion.UnifyResult
     {
         /// <summary>
         /// 设置返回 200 状态码列表
-        /// <para>默认：401，403，如果设置为 null，则标识所有状态码都返回 200 </para>
+        /// <para>默认：401，403，404，如果设置为 null，则标识所有状态码都返回 200 </para>
         /// </summary>
-        public int[] Return200StatusCodes { get; set; } = new[] { 401, 403 };
+        public int[] Return200StatusCodes { get; set; } = new[] { 404, 401, 403 };
 
         /// <summary>
         /// 适配（篡改）Http 状态码（只支持短路状态码，比如 401，403，404，500 等）